+++ NEWS.md
@@ -1,3 +1,7 @@
+# modeltime.gluonts 0.2.0.9000
+
+Improvements made to connect with the GluonTS Python Environment on Startup. 
+
 # modeltime.gluonts 0.2.0
 
 ### Documentation